requirements.txt
streamlit
numpy
pandas
scikit-learn==1.6.1
joblib
nltk
plotly
boto3
python-dotenv
tensorflow
contractions
-e .